tweak tests
