+++ test/plots/index.js
@@ -28,7 +28,7 @@ export {default as binTimestamps} from "./bin-timestamps.js";
 export {default as boxplot} from "./boxplot.js";
 export {default as caltrain} from "./caltrain.js";
 export {default as caltrainDirection} from "./caltrain-direction.js";
-export {default as carsJiggle} from "./cars-jiggle.js";
+export {default as carsJitter} from "./cars-jitter.js";
 export {default as carsMpg} from "./cars-mpg.js";
 export {default as carsParcoords} from "./cars-parcoords.js";
 export {default as clamp} from "./clamp.js";
